refactor: remove legacy setting.py entry point
- Remove the old setting.py file after migrating all logic to modular structure
- Update imports from setting.py to settings module in critical paths
- Update test files to use new modular settings imports
- All core functionality maintained through settings/ package
- Essential tests pass confirming backward compatibility

This removes the old centralized entry point in favor of the
new modular settings architecture with organized submodules:
- settings/constants.py - for constants and mappings
- settings/path_manager.py - for path and folder management
- settings/model_utils.py - for model-related utilities
- settings/initialization.py - for initialization routines
- settings/config_manager.py - for configuration management

Remaining non-critical imports of setting.py can be updated
incrementally as those modules are touched in future work.

Resolves the entry point migration task from the modular refactor.
