+++ scripts/civitai_manager_libs/settings/initialization.py
@@ -1,0 +1,56 @@
+"""Initialization utilities for the application."""
+
+import os
+
+from ..logging_config import get_logger
+from ..ui.notification_service import GradioNotificationService, set_notification_service
+from .constants import SC_DATA_ROOT
+from .path_manager import init_paths, migrate_existing_files, load_model_folder_data
+
+logger = get_logger(__name__)
+
+# Compatibility layer variables
+_compat_layer = None
+
+
+def set_compatibility_layer(compat_layer):
+    """Set compatibility layer (called by main program)."""
+    global _compat_layer
+    logger.info("Setting compatibility layer.")
+    _compat_layer = compat_layer
+    from .path_manager import initialize_extension_base
+
+    initialize_extension_base()
+
+
+def init():
+    """Initialize application with notification service setup."""
+    from . import config_manager
+
+    set_notification_service(GradioNotificationService())
+    from .path_manager import get_extension_base
+
+    logger.info(f"Initializing with extension_base={get_extension_base()}")
+    try:
+        os.makedirs(SC_DATA_ROOT, exist_ok=True)
+    except Exception as e:
+        logger.warning(f"Failed to create SC_DATA_ROOT for migration: {e}")
+    migrate_existing_files()
+    init_paths(config_manager)
+    config_manager.load_settings()
+    load_model_folder_data(config_manager)
+
+
+def set_NSFW(enable, level="None"):
+    """Sets NSFW filtering options."""
+    from . import config_manager
+
+    config_manager.set_setting('NSFW_filtering_enable', enable)
+    config_manager.set_setting('NSFW_level_user', level)
+
+
+def save_NSFW():
+    """Save NSFW settings to the configuration file."""
+    from . import config_manager
+
+    return config_manager.save_settings()
