+++ scripts/civitai_manager_libs/settings/path_manager.py
@@ -1,0 +1,196 @@
+"""Path management utilities for the application."""
+
+import os
+import shutil
+
+from ..compat.compat_layer import CompatibilityLayer
+from ..conditional_imports import import_manager
+from ..logging_config import get_logger
+from .constants import SC_DATA_ROOT, default_model_folders, preview_image_ext
+
+logger = get_logger(__name__)
+
+# Global variables
+root_path = os.getcwd()
+extension_base = ""
+model_folders = default_model_folders.copy()
+
+# File paths for various application data
+shortcut = os.path.join(SC_DATA_ROOT, "CivitaiShortCut.json")
+shortcut_setting = os.path.join(SC_DATA_ROOT, "CivitaiShortCutSetting.json")
+shortcut_classification = os.path.join(SC_DATA_ROOT, "CivitaiShortCutClassification.json")
+shortcut_civitai_internet_shortcut_url = os.path.join(SC_DATA_ROOT, "CivitaiShortCutBackupUrl.json")
+shortcut_recipe = os.path.join(SC_DATA_ROOT, "CivitaiShortCutRecipeCollection.json")
+
+shortcut_thumbnail_folder = os.path.join(SC_DATA_ROOT, "sc_thumb_images")
+shortcut_recipe_folder = os.path.join(SC_DATA_ROOT, "sc_recipes")
+shortcut_info_folder = os.path.join(SC_DATA_ROOT, "sc_infos")
+shortcut_gallery_folder = os.path.join(SC_DATA_ROOT, "sc_gallery")
+
+
+def get_extension_base():
+    """Get the extension base path."""
+    return extension_base
+
+
+def set_extension_base(path):
+    """Set the extension base path."""
+    global extension_base
+    extension_base = path
+
+
+def initialize_extension_base():
+    """Initialize extension base path."""
+    global extension_base
+    logger.debug("Initializing extension base path.")
+    compat = CompatibilityLayer.get_compatibility_layer()
+    if compat and hasattr(compat, 'path_manager'):
+        extension_base = str(compat.path_manager.get_extension_path())
+        logger.debug(f"Set extension_base from compat.path_manager: {extension_base}")
+    else:
+        current_dir = os.path.dirname(os.path.abspath(__file__))
+        extension_base = os.path.dirname(os.path.dirname(os.path.dirname(current_dir)))
+        logger.debug(f"Fallback extension_base: {extension_base}")
+
+
+def get_no_card_preview_image():
+    """Get path to no card preview image."""
+    return os.path.join(extension_base, "img", "card-no-preview.png")
+
+
+def get_nsfw_disable_image():
+    """Get path to NSFW disable image."""
+    return os.path.join(extension_base, "img", "nsfw-no-preview.png")
+
+
+def init_paths(config_manager):
+    """Initialize and create necessary directories."""
+    from .constants import SD_DATA_ROOT
+
+    download_images_folder = config_manager.get_setting('download_images_folder')
+    dirs = [
+        SC_DATA_ROOT,
+        SD_DATA_ROOT,
+        os.path.join(SD_DATA_ROOT, "models"),
+        os.path.join(SD_DATA_ROOT, "output"),
+        download_images_folder,
+        shortcut_thumbnail_folder,
+        shortcut_recipe_folder,
+        shortcut_info_folder,
+        shortcut_gallery_folder,
+    ]
+    for d in dirs:
+        if d:  # Only create if directory is not None/empty
+            try:
+                os.makedirs(d, exist_ok=True)
+            except Exception as e:
+                logger.warning(f"Failed to create directory {d}: {e}")
+
+
+def migrate_existing_files():
+    """Migrate existing files and folders in root to new data_sc structure."""
+    mapping = {
+        "CivitaiShortCut.json": shortcut,
+        "CivitaiShortCutSetting.json": shortcut_setting,
+        "CivitaiShortCutClassification.json": shortcut_classification,
+        "CivitaiShortCutRecipeCollection.json": shortcut_recipe,
+        "CivitaiShortCutBackupUrl.json": shortcut_civitai_internet_shortcut_url,
+        "sc_gallery": shortcut_gallery_folder,
+        "sc_thumb_images": shortcut_thumbnail_folder,
+        "sc_infos": shortcut_info_folder,
+        "sc_recipes": shortcut_recipe_folder,
+    }
+    for old, new in mapping.items():
+        if os.path.exists(old) and not os.path.exists(new):
+            try:
+                shutil.move(old, new)
+                logger.info(f"Moved {old} to {new}")
+            except Exception as e:
+                logger.warning(f"Failed to move {old} to {new}: {e}")
+
+    for entry in os.listdir('.'):
+        if entry.startswith('sc_'):
+            old = entry
+            new = os.path.join(SC_DATA_ROOT, entry)
+            if os.path.exists(old) and not os.path.exists(new):
+                try:
+                    shutil.move(old, new)
+                    logger.info(f"Moved {old} to {new}")
+                except Exception as e:
+                    logger.warning(f"Failed to move {old} to {new}: {e}")
+
+
+def load_model_folder_data(config_manager):
+    """Load configuration data and update model folders."""
+    logger.info("Loading model folder configuration data.")
+    global model_folders
+
+    compat = CompatibilityLayer.get_compatibility_layer()
+    if compat and hasattr(compat, 'path_manager'):
+        logger.debug("Using compat.path_manager for model folders.")
+        for model_type, path_key in {
+            'TextualInversion': 'embeddings',
+            'Hypernetwork': 'hypernetworks',
+            'Checkpoint': 'Stable-diffusion',
+            'LORA': 'Lora',
+        }.items():
+            model_path = compat.path_manager.get_model_path(path_key)
+            if model_path:
+                model_folders[model_type] = model_path
+                logger.debug(f"Set {model_type} folder: {model_path}")
+    else:
+        logger.debug("Fallback to import_manager for model folders.")
+        shared = import_manager.get_webui_module('shared')
+        if shared and hasattr(shared, 'cmd_opts'):
+            cmd_opts = shared.cmd_opts
+            for model_type, attr in {
+                'TextualInversion': 'embeddings_dir',
+                'Hypernetwork': 'hypernetwork_dir',
+                'Checkpoint': 'ckpt_dir',
+                'LORA': 'lora_dir',
+            }.items():
+                if hasattr(cmd_opts, attr) and getattr(cmd_opts, attr):
+                    model_folders[model_type] = getattr(cmd_opts, attr)
+                    logger.debug(f"Set {model_type} folder (fallback): {getattr(cmd_opts, attr)}")
+
+    environment = config_manager.load_settings()
+    if not environment:
+        return
+
+    logger.info(f"Loaded environment: {environment}")
+
+    user_folders = environment.get('model_folders', {})
+    for folder_key in ['LoCon', 'Wildcards', 'Controlnet', 'AestheticGradient', 'Poses', 'Other']:
+        if folder_key in user_folders:
+            model_folders[folder_key] = user_folders[folder_key]
+            logger.debug(f"Set {folder_key} folder: {user_folders[folder_key]}")
+
+
+def get_model_folders():
+    """Returns a list of all model folder paths."""
+    return list(model_folders.values())
+
+
+def get_image_url_to_shortcut_file(modelid, versionid, image_url):
+    """Generates the local file path for a shortcut image from a URL."""
+    if image_url:
+        version_image_prefix = f"{versionid}-"
+        model_path = os.path.join(shortcut_info_folder, str(modelid))
+        image_id, _ = os.path.splitext(os.path.basename(image_url))
+        return os.path.join(model_path, f"{version_image_prefix}{image_id}{preview_image_ext}")
+    return None
+
+
+def get_image_url_to_gallery_file(image_url):
+    """Generates the local file path for a gallery image from a URL."""
+    if image_url:
+        image_id, _ = os.path.splitext(os.path.basename(image_url))
+        return os.path.join(shortcut_gallery_folder, f"{image_id}{preview_image_ext}")
+    return None
+
+
+# Initialize extension_base on module load
+try:
+    initialize_extension_base()
+except Exception:
+    pass
