+++ scripts/civitai_manager_libs/settings/setting_categories.py
@@ -1,5 +1,6 @@
 """Manages setting categories and their definitions."""
 
+
 class SettingCategories:
     """Defines and manages setting categories."""
