+++ tests/test_recipe_actions/test_recipe_management.py
@@ -1,7 +1,7 @@
 import pytest
 from unittest.mock import patch
 
-import scripts.civitai_manager_libs.setting as setting
+from scripts.civitai_manager_libs import settings
 from scripts.civitai_manager_libs.recipe_actions.recipe_management import RecipeManager
 from scripts.civitai_manager_libs.exceptions import ValidationError, FileOperationError
 
@@ -11,10 +11,10 @@ def use_tmp_recipe_file(tmp_path, monkeypatch):
     """Redirect recipe storage to a temporary file for tests."""
     tmp_file = tmp_path / "recipes.json"
     tmp_folder = tmp_path / "sc_recipes"
-    monkeypatch.setattr(setting, "shortcut_recipe", str(tmp_file))
-    monkeypatch.setattr(setting, "shortcut_recipe_folder", str(tmp_folder))
-    monkeypatch.setattr(setting, "NEWRECIPE", "New Recipe")
-    monkeypatch.setattr(setting, "PLACEHOLDER", "Select Classification")
+    monkeypatch.setattr(settings, "shortcut_recipe", str(tmp_file))
+    monkeypatch.setattr(settings, "shortcut_recipe_folder", str(tmp_folder))
+    monkeypatch.setattr(settings, "NEWRECIPE", "New Recipe")
+    monkeypatch.setattr(settings, "PLACEHOLDER", "Select Classification")
     yield
