+++ tests/test_shortcut_search_filter.py
@@ -47,7 +47,7 @@ def test_get_shortcuts_list(search_filter):
 
 
 def test_type_filter(search_filter, monkeypatch):
-    import scripts.civitai_manager_libs.setting as setting
+    from scripts.civitai_manager_libs import settings
 
     monkeypatch.setitem(config_manager.get_setting('ui_typenames'), 'X', 'typeA')
     names = search_filter.get_shortcuts_list(['X'])
